Move LootrunFeature to LootrunFunctionality
