+++ common/src/main/java/com/wynntils/functionalities/LootrunFunctionality.java
@@ -2,10 +2,9 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.features;
+package com.wynntils.functionalities;
 
-import com.google.common.collect.ImmutableList;
-import com.wynntils.core.features.FeatureBase;
+import com.wynntils.core.functionalities.FunctionalityBase;
 import com.wynntils.mc.event.ClientTickEvent;
 import com.wynntils.mc.event.PlayerInteractEvent;
 import com.wynntils.mc.event.RenderLevelLastEvent;
@@ -16,26 +15,16 @@ import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-public class LootrunFeature extends FeatureBase {
-
-    public LootrunFeature() {
-        setupEventListener();
-    }
-
-    @Override
-    protected void onInit(ImmutableList.Builder<Condition> conditions) {
-        LootrunUtils.LOOTRUNS.mkdirs();
-    }
-
+public class LootrunFunctionality extends FunctionalityBase {
     @SubscribeEvent
-    public void recordMovement(ClientTickEvent event) {
+    public static void recordMovement(ClientTickEvent event) {
         if (event.getTickPhase() == ClientTickEvent.Phase.START) {
             LootrunUtils.recordMovementIfRecording();
         }
     }
 
     @SubscribeEvent
-    public void onRightClickBlock(PlayerInteractEvent.RightClickBlock event) {
+    public static void onRightClickBlock(PlayerInteractEvent.RightClickBlock event) {
         BlockState block = event.getWorld().getBlockState(event.getPos());
         if (block.is(Blocks.CHEST)) {
             LootrunUtils.setLastChestIfRecording(event.getPos());
@@ -43,14 +32,14 @@ public class LootrunFeature extends FeatureBase {
     }
 
     @SubscribeEvent
-    public void onOpen(ScreenOpenedEvent event) {
+    public static void onOpen(ScreenOpenedEvent event) {
         if (ContainerUtils.isLootChest(event.getScreen())) {
             LootrunUtils.addChestIfRecording();
         }
     }
 
     @SubscribeEvent
-    public void onRenderLastLevel(RenderLevelLastEvent event) {
+    public static void onRenderLastLevel(RenderLevelLastEvent event) {
         LootrunUtils.render(event.getPoseStack());
     }
 }
